### 2023-09-05 #5807 @editorialbot
[PRE REVIEW]: Cleaving: a LAMMPS package to compute surface free energies
**Submitting author:** @demonico85 (<a href="http://orcid.org/0000-0001-5676-8527">Nicodemo Di Pasquale</a>)
**Repository:** https://github.com/demonico85/cleaving
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0.0
