### 2023-09-05 #5807 @Kevin-Mattheus-Moerman
@demonico85 I am the AEiC for this track and here to help with initial steps.

- [ ] It looks like the license was not detected on your `joss` branch. We can work around this but it may be easier to make sure that branch is complete too.
- [ ] Can you check the potentially missing DOIs above ☝️ ? You can amend your .bib file and update the paper by calling:
`@editorialbot generate pdf`
and you can then check the references again by calling:
`@editorialbot check references`
