+++ _sources/fix_well.md.txt
@@ -20,11 +20,11 @@ On each side of the cleaving plane we place two planes with fixed interaction si
 The well potential we use in this work is  defined as {footcite:t}`handel2008direct`: 
 
 $$
-\psi(r;\lambda) = 
-	\begin{cases}
-		\lambda dw\left[\left(\frac{r}{rw}\right)^2-1\right]^{P} \:\:\: r < rw \\
-		0  \:\:\: r \geq rw
-	\end{cases} 
+	\psi(r;\lambda) = 
+		\begin{cases}
+			\lambda d_w\left[\left(\frac{r}{r_w}\right)^2-1\right]^{P} \:\:\: r < r_w \\
+			0  \:\:\: r \geq r_w
+		\end{cases} 
 $$
 where $r=|\mathbf{r}_{atoms}-\mathbf{r}_{wells}|$ for each couple atom-well.
 
