### 2023-09-06 a65a92a
Update the bib file to include DOIs
Fix the problems raised
[here](https://github.com/openjournals/joss-reviews/issues/5807)
