+++ joss/paper.bib
@@ -6,7 +6,8 @@
   number={16},
   pages={7651--7657},
   year={2003},
-  publisher={American Institute of Physics}
+  publisher={American Institute of Physics},
+  doi={10.1063/1.1563248},
 }
 @article{Thompson2022,
   title={LAMMPS-a flexible simulation tool for particle-based materials modeling at the atomic, meso, and continuum scales},
@@ -15,7 +16,8 @@
   volume={271},
   pages={108171},
   year={2022},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.cpc.2021.108171}
 }
 @article{DiPasquale2022cleaving,
   title={Cleaving Method for Molecular Crystals and Its Application to Calculation of the Surface Free Energy of Crystalline $\beta$-d-Mannitol at Room Temperature},
@@ -35,14 +37,16 @@
   number={15},
   pages={154705},
   year={2020},
-  publisher={AIP Publishing LLC}
+  publisher={AIP Publishing LLC},
+  doi={10.1063/5.0028219}
 }
 
 @techreport{GibbsCollectedWorks,
   title={The collected works of J. Willard Gibbs.},
   author={Gibbs, J. W.},
   year={1948},
-  institution={Yale Univ. Press,}
+  institution={Yale Univ. Press},
+  doi={10.1038/124119a0}
 }
 
 @article{Broughton1983,
@@ -55,7 +59,8 @@
 	publisher = {AIP},
 	title = {Molecular dynamics investigation of the crystal--fluid interface. I. Bulk properties},
 	volume = {79},
-	year = {1983}
+	year = {1983},
+	doi={10.1063/1.445633}
 }
 
 @article{Broughton1986Cleaving,
@@ -66,7 +71,8 @@
   number={10},
   pages={5759--5768},
   year={1986},
-  publisher={American Institute of Physics}
+  publisher={American Institute of Physics},
+  doi={10.1063/1.449884}
 }
 
 @article{Davidchack00prl,
@@ -88,7 +94,8 @@
   number={8},
   pages={086102},
   year={2005},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/physrevlett.94.086102}
 }
 
 @article{Davidchack03direct,
@@ -99,7 +106,8 @@
   number={16},
   pages={7651--7657},
   year={2003},
-  publisher={American Institute of Physics}
+  publisher={American Institute of Physics},
+  doi={10.1063/1.1563248}
 }
 
 @article{Handel08prl,
@@ -110,7 +118,8 @@
   number={3},
   pages={036104},
   year={2008},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/physrevlett.100.036104}
 }
 
 @article{Davidchack12ice,
@@ -131,7 +140,8 @@
   volume={155},
   pages={136--143},
   year={2018},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.commatsci.2018.08.035}
 }
 
 @article{Liu2013csm,
@@ -141,7 +151,8 @@
   volume={74},
   pages={92--100},
   year={2013},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.commatsci.2013.03.018}
 }
 
 
@@ -153,7 +164,8 @@
   number={19},
   pages={194108},
   year={2016},
-  publisher={AIP Publishing LLC}
+  publisher={AIP Publishing LLC},
+  doi={10.1063/1.4967521}
 }
 
 @article{addula2020computation,
@@ -164,5 +176,6 @@
   number={15},
   pages={154504},
   year={2020},
-  publisher={AIP Publishing LLC}
+  publisher={AIP Publishing LLC},
+  doi={10.1063/5.0028653}
 }
