+++ joss/paper.bib
@@ -27,7 +27,8 @@
   number={13},
   pages={2134--2141},
   year={2022},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
+  doi={10.1021/acs.jpca.2c00604}
 }
 @article{DiPasquale2020shuttleworth,
   title={Shuttleworth equation: A molecular simulations perspective},
@@ -83,7 +84,8 @@
   number={22},
   pages={4751},
   year={2000},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/PhysRevLett.85.4751}
 }
 
 @article{Davidchack05prl,
@@ -130,7 +132,8 @@
   number={7},
   pages={2383--2390},
   year={2012},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
+  doi={10.1021/ct300193e}
 }
 
 @article{Yan18structural,
