### 2023-09-06 26776a4
added separate switching lj and coul for tip4p pair style
