### 2023-09-08 8e36eb9
fixed storing LJ and C in tip4p_cleav_cut
