### 2023-09-12 #5807 @demonico85
@Kevin-Mattheus-Moerman

We have fixed the problem with the references and the license.

We also identified two possible reviewers
@ptmerz
@corettialessandro

@lorenzo-rovigatti  is another author of the submitted paper and is participating to the review process
