### 2023-09-20 #5807 @mbarzegary
Hi @ptmerz @corettialessandro @anilkunwar 👋
Would any of you be willing to review this submission for [JOSS](https://joss.theoj.org/about)? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Paper: https://raw.githubusercontent.com/openjournals/joss-papers/joss.05807/joss.05807/10.21105.joss.05807.pdf
Software: https://github.com/demonico85/cleaving
