### 2023-10-09 demonico85/cleaving#6 @alexgorfer
[BUG] cleav_test.sh error; Atoms have moved too far apart for minimum image
Hey @demonico85,
After executing cleav_test.sh, during step 1, the loop crashes at Wall position 0.9 with the output:

```
System init for write_data ...
Generated 0 of 0 mixed pair_coeff terms from geometric mixing rule
Wall Position 0.9
Generated 0 of 0 mixed pair_coeff terms from geometric mixing rule
Setting up Verlet run ...
  Unit style    : lj
  Current step  : 240
  Time step     : 0.005
ERROR on proc 2: Atoms have moved too far apart (-8.926507890180693e+245) for minimum image
 (../domain.cpp:986)
Last command: run   ${eqnts}
ERROR on proc 3: Atoms have moved too far apart (-8.926507890180693e+245) for minimum image
 (../domain.cpp:986)
Last command: run   ${eqnts}
ERROR on proc 0: Atoms have moved too far apart (-8.926507890180693e+245) for minimum image
 (../domain.cpp:986)
Last command: run   ${eqnts}
ERROR on proc 1: Atoms have moved too far apart (-8.926507890180693e+245) for minimum image
 (../domain.cpp:986)
Last command: run   ${eqnts}`
```

It always crashes at 0.9, however sometimes in /data it generates up to Fstep1.24.data, sometimes only to Fstep.1.12.data

It looks like some overflow - Do you know what the issue might be?

**Desktop:**
 - OS: Linux
 - Cleaving Version Sep 6 2023
-  lammps-2Aug2023 built with gnu compilers.
