### 2023-10-09 #5886 @ptmerz
Confirming that the test does not work for me either, I added a note at the issue above.
