### 2023-10-09 #5886 @ptmerz
I am also wondering if you could comment on your choice of distribution / installation method. Why did you decide against contributing your work to LAMMPS directly? I think that would have many benefits, including a review from the LAMMPS maintainers, reducing the chance of integration problems (for the current or a future version), as well as a simpler installation process for users. Downloading a third-party repo and copying source files into the LAMMPS distribution certainly works, but isn't the greatest user experience, and might be especially challenging when trying to use the package on HPC resources.
