### 2023-10-09 #5886 @ptmerz
Another point Re: "Contribution and authorship: Has the submitting author (@demonico85) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?". It's easy to verify through the git log that @demonico85 and @lorenzo-rovigatti (first and last author) have contributed significantly to the software. The git log does not allow to verify Ruslan Davidchacks contribution. Can you comment on including them as an author?
