### 2023-10-09 #5886 @ptmerz
Further points Re: "Documentation" in the checklist above:

* **A statement of need:** The documentation does not contain a clear statement of need. I'll note that the paper contains an excellent statement of need, but the documentation doesn't. I'd suggest to add a shortened version of the statement of need of the paper to the front page of the documentation.
* **Automated tests:** In my understanding, the test which is described in the documentation merely tests whether the commands run without error. Can you explain whether that does also represent a test of _correctness_? Could you, for example, use a regression test to ensure that the resulting values are actually correct? It also doesn't seem like there are any unit tests checking the correctness of the different parts of the code.
* **Community guidelines:** The repo and documentation do not contain community guidelines.

I'd like to note that the documentation of the different LAMMPS commands is excellent. I really appreciate the work you put into explaining each command in depth, often including formulas, references and plots.
