### 2023-10-09 #5886 @ptmerz
Here's a few more things I noticed:

- **Naming:** The paper uses Cleaving (in the title) and CLEAVING (line 77 / 80), the documentation uses Cleaving and USER-CLEAVING, where as the name next to the home button (blue area top right, above the search bar) is "SIFT".
- **Paper:** Just a small suggestion here, but I feel that the explanation of where the "cleaving" name comes from (line 99) would rather fit in the method description (maybe around lines 52 or 58) than in the functionality of the package.
- **utils:** You have a large number of files in the `utils/` folder, which make up a non-trivial part of your source code. Unfortunately, most of these are not documented. I see that 3 of them are described in the last step of the examples, but one of them (`calcSFE.m`) isn't even present in my checkout. This leaves 40 files without any description. Do you consider these files part of this submission? In this case, I think these would need a description in the documentation, and ideally tests.

I hope that all the points that I raised above don't come over as if I was thinking negatively of the submission. In fact, I think that the work is highly relevant, the LAMMPS functions are nicely documented, and the paper is very well written.
