### 2023-10-10 ee4504d
Update fix_move_duplicate.md
