+++ _sources/pair_BGs3lam.md.txt
@@ -28,13 +28,14 @@ cutoff1 = global internal cut-off
 cutoff2 = global external cut-off
 ```
 
-the arguments in the `pair_coeff` are all optional except for the atom types. Note that the order is important. If we want to specify a new lambda for a different pair of types we can just write
+The arguments in the `pair_coeff` are all optional except for the atom types. However, if we need to specify a particular argument, we need to specify also all the preceeding ones. 
+If we want to specify a new lambda for a different pair of types we can just write
 
 ```
 pair_coeff 1 1 1.0
 ```
 
-However, in order to specify a new (e.g., external) cutoff we need to rewrite all the preceeding arguments:
+Instead, in order to specify a new (e.g., external) cutoff we need to rewrite all the preceeding arguments:
 
 ```
 pair_coeff 1 1 lambda Dfac cutoff1 cutoff2
@@ -51,7 +52,7 @@ pair_coeff 1 3 -1.0
 
 ## Description
 
-This pair style implements the Broughton and Gilmer modification to Lennard-Jones potential {footcite:t}`Broughton1983` (see [pair lj/BG]{pairBG.md}) to be used in the step3 of the cleaving algorithm. This pair style returns also an array with all the calculated interactions which are needed to calculate the work in the step3. This array can be accessed using the new compute [compute paircleav]}{compute_pcleav.md}. 
+This pair style implements the {footcite:t}`Broughton1983` modification to Lennard-Jones potential  (see [pair lj/BG]{pairBG.md}) to be used in the step3 of the cleaving algorithm. This pair style returns also an array with all the calculated interactions which are needed to calculate the work in the step3. This array can be accessed using the new compute [compute paircleav]}{compute_pcleav.md}. 
 
 
 The potential implemented in this pair style is 
@@ -65,11 +66,10 @@ $$
 		\end{cases}
 $$
 
-where $r_{ln}=|\mathbf{r}_l-\mathbf{r}_n|$ for each couple of atoms $l,n$ in the system, and $C_1, C_2, C_3, C_4, C_5$ are constants we used the values reported in {footcite:t}`davidchack2003direct`.
-The constants are hardcorded within the pair style and they not need to be defined.
+where $r_{ln}=|\mathbf{r}_l-\mathbf{r}_n|$ for each couple of atoms $l,n$ in the system, and $C_1, C_2, C_3, C_4, C_5$ are constants hardcorded within the pair style, therefore they not need to be defined. For C_1, C_2, C_3, C_4, C_5$ we used the values reported in {footcite:t}`davidchack2003direct`. 
 
 ````{note}
-* $lambda$ must vary in the (closed) interval [0,1]
+* $\lambda$ must vary in the (closed) interval [0,1]
 * There is no internal check that $\lambda$ is changing consistently (i.e., always decreasing or increasing)
 ````
    
@@ -84,7 +84,7 @@ $$
 		\end{cases}
 $$
 
-In this case the the expression of the potential is multiplied by the derivatibe of $\lambda$ with respect itself which is equal to 1 (i.e., $\partial \lambda / \partial \lambda=1$), and therefore the value of Dfrac must be set equal to 1.
+In this case, the the expression of the potential is multiplied by the derivative of $\lambda$ with respect itself, which is equal to 1 (i.e., $\partial \lambda / \partial \lambda=1$). Therefore, the value of Dfrac must be set equal to 1.
 
 Using this style we can represent different switches between different states of the system:
 
@@ -135,7 +135,7 @@ $$
 $$  
 
 for $\lambda \in [0,1]$, where we dropped the dependence on $\mathbf{p}$ and $\mathbf{q}$ for simplicity.
[…]
