### 2023-10-10 a214df0
Update pair_BGs3Nlam.md
