+++ _sources/fix_move_duplicate.md.txt
@@ -14,23 +14,19 @@ fix ID group-ID move/dupl C_ID
 This fix allows to translate the duplicate atoms of the same distance crossed by the corresponding real atoms. In step3 of the cleaving methodology the duplicate atoms allow to calculate the cross-cleaving plane interactions but they need to "follow" the real atoms when they move in a new position each time step. 
 
 
-In the step3 the system (solid+liquid) is duplicated and `real` and `duplicate` atom types are defined (see [example LJ-SL](example_SL_walls.md) for more details).
+In the step3 the system (solid+liquid) is duplicated and `real` and `duplicate` atom types are defined (see [example LJ-SL](example_SL_walls.md) for more details) as showin in the next Figure.
 
 
 ![definition](../figs/dupl1.png "Definition of duplicate atoms")
 
 
-The atoms interacts with one another. The duplicate atoms are used to keep track of the interactions crossing the cleaving planes, which are the ones to be calculated in order to find the total work in step3. 
-
-The equation of motion must be integrated _only_ for the real atoms. The atoms then move by a distance $\Delta \mathbf{x}$.
+The duplicate atoms are used to keep track of the interactions crossing the cleaving planes, which are the ones to be calculated in order to find the total work in step3. 
+However, the equation of motion must be integrated _only_ for the real atoms. After the integration steo the _real_ atoms move by a distance $\Delta \mathbf{x}$ (see next Figure).
 
 ![move](../figs/dupl2.png "Move the real atoms")
 
 
-This fix allows the duplicated atoms to follow the movement of the real atoms. By specifying this fix we are moving the atoms of the same quantity as the real atoms.
-
-
-The equation of motion must be integrated _only_ for the real atoms. The atoms then move by a distance $\Delta \mathbf{x}$.
+This fix allows the duplicated atoms to follow the movement of the corresponding real atoms. By specifying this fix we are moving the _duplicated_ atoms by the same quantity calculated for the real atoms as shown in the next Figure.
 
 ![move](../figs/dupl3.png "Move the real atoms")
 
