### 2023-10-10 8a233ac
Update pair_BGs3Nlam.md
