### 2023-10-10 9341f2e
Update pair_BGs3lam.md
