### 2023-10-10 f931ec8
Update the paper as requested in https://github.com/openjournals/joss-reviews/issues/5886
