### 2023-10-10 dd51a17
Strengthen the test script (and check that #6 is fixed)
