### 2023-10-11 2bd3bd2
Fix the documentation as per the JOSS review (https://github.com/openjournals/joss-reviews/issues/5886)
