+++ docs/source/refs.bib
@@ -9,6 +9,89 @@
   publisher={American Institute of Physics}
 }
 
+@article{Davidchack00prl,
+  title={Direct calculation of the hard-sphere crystal/melt interfacial free energy},
+  author={Davidchack, R. L. and Laird, B. B.},
+  journal={Physical review letters},
+  volume={85},
+  number={22},
+  pages={4751},
+  year={2000},
+  publisher={APS},
+  doi={10.1103/PhysRevLett.85.4751}
+}
+
+@article{Davidchack05prl,
+  title={Crystal structure and interaction dependence of the crystal-melt interfacial free energy},
+  author={Davidchack, Ruslan L and Laird, Brian B},
+  journal={Physical review letters},
+  volume={94},
+  number={8},
+  pages={086102},
+  year={2005},
+  publisher={APS},
+  doi={10.1103/physrevlett.94.086102}
+}
+
+@article{Davidchack12ice,
+  title={Ice Ih--water interfacial free energy of simple water models with full electrostatic interactions},
+  author={Davidchack, Ruslan L and Handel, Richard and Anwar, Jamshed and Brukhno, Andrey V},
+  journal={Journal of Chemical Theory and Computation},
+  volume={8},
+  number={7},
+  pages={2383--2390},
+  year={2012},
+  publisher={ACS Publications},
+  doi={10.1021/ct300193e}
+}
+
+@article{Handel08prl,
+  title={Direct calculation of solid-liquid interfacial free energy for molecular systems: TIP4P ice-water interface},
+  author={Handel, Richard and Davidchack, Ruslan L and Anwar, Jamshed and Brukhno, Andrey},
+  journal={Physical review letters},
+  volume={100},
+  number={3},
+  pages={036104},
+  year={2008},
+  publisher={APS},
+  doi={10.1103/physrevlett.100.036104}
+}
+
+@article{DiPasquale2022cleaving,
+  title={Cleaving Method for Molecular Crystals and Its Application to Calculation of the Surface Free Energy of Crystalline $\beta$-d-Mannitol at Room Temperature},
+  author={Di Pasquale, N. and Davidchack, R. L.},
+  journal={The Journal of Physical Chemistry A},
+  volume={126},
+  number={13},
+  pages={2134--2141},
+  year={2022},
+  publisher={ACS Publications},
+  doi={10.1021/acs.jpca.2c00604}
+}
+
+@article{addula2020computation,
+  title={Computation of solid--fluid interfacial free energy in molecular systems using thermodynamic integration},
+  author={Addula, Ravi Kumar Reddy and Punnathanam, Sudeep N},
+  journal={The Journal of Chemical Physics},
+  volume={153},
+  number={15},
+  pages={154504},
+  year={2020},
+  publisher={AIP Publishing LLC},
+  doi={10.1063/5.0028653}
+}
+
+@article{Liu2013csm,
+  title={Molecular dynamics calculation of solid--liquid interfacial free energy and its anisotropy during iron solidification},
+  author={Liu, J and Davidchack, RL and Dong, HB},
+  journal={Computational materials science},
+  volume={74},
+  pages={92--100},
+  year={2013},
+  publisher={Elsevier},
+  doi={10.1016/j.commatsci.2013.03.018}
+}
+
 @article{Broughton1983,
 	author = {Broughton, J. Q. and Gilmer, G. H.},
 	date-added = {2019-03-26 17:55:17 +0000},
@@ -20,6 +103,18 @@
 	title = {Molecular dynamics investigation of the crystal--fluid interface. I. Bulk properties},
 	volume = {79},
 	year = {1983}
+}
+
+@article{Broughton1986Cleaving,
[…]
