### 2023-10-11 17b7de3
Fix a bibliography entry
