### 2023-10-11 #5886 @lorenzo-rovigatti
Thanks a lot for your thorough review, @ptmerz !
We are going through all the points you have raised. For now we have fixed the paper, added a statement of need to the docs (in the form of a "background" section) and enforced using CLEAVING throughout the docs.
I'll update the following list as soon as we are done with each item:

* [x] A statement of need
* [ ] Automated tests
* [ ] Community guidelines
* [x] Naming
* [x] Paper
* [ ] utils

We will provide answer (and fixes) to the issues created in the repo directly there!
