### 2023-10-11 demonico85/cleaving#6 @demonico85
We have fixed this bug, now the test should run as intended.

Tested on lammps-2Aug2023 built with gnu compilers
