### 2023-10-12 e10797f
Add a "contributing" section to the README (as requested in the [JOSS review](https://github.com/openjournals/joss-reviews/issues/5886))
