### 2023-10-12 #5886 @alexgorfer
The work implements an accessible cleaving method in LAMMPS which will certainly be of interest for the readership of JOSS. Overall the work is done carefully, the documentation is clearly written and figures are informative. Adding to what has been put forth ptmerz already, here are some additional comments:

**Functionality** The main point that I would like discussed in more depth is the range of applicability of the software. What I mean is that in the documentation and paper we learn about prior uses of the method including hard and soft-spheres, LJ, TIP4P, EAM and others and that the adoption for different forcefields requires a more convenient simulation interface. Which is the motivation for this work. To me this suggested that the software can already cleave TIP4P, EAM and others, but I am not sure if this is the case since in TIP4P, Handel and Davichack et al. developed an orientational cleaving potential for step 2, and in the case of EAM, Liu and coworkers added a soft repuslive part to the potential in order to compensate their calculation at non-coexistence, again to achieve crystal layers in step 2. Since I do not see this implemented, could you clarify in the documentation and paper what the current implementation is capable of? Say I want to repeat Handel et al.s work, can I:
1.  use existing pair_styles and fixes (it would be great to have an example in this case)
2.  or do I need edit wellPforce to include orientation (add some documentation to use the interface for my custom force-fields)
3.  or can I not use this package for this purpose yet (then state this more clearly)
If the latter is the case and its not possible to do step 2 for arbitrary systems, is it still possible to do step 1,3 and 4? So the solid-vacuum TIP4P interface. TIP4P can be made with lj/cut/coul and fix rigid and if that all works with your implementation it would be awesome to add an example of this more complicated scenario.

#

Minor points:
- In evaluating step3 of examples/lj_SV using a.out I got ERROR file zdir.dat not found. I copied zdir.dat into /out/ and it worked. All other steps in the example worked without issue!
- Does it have to be Matlab to calculate the integrals? Something non-proprietary would be preferable, like the calcintegrals.f90 in utils.
- Speaking of utils, there is no documentation for these scripts but some sound very useful for users.
- The citation style on line 68-73 is not uniform (author, year) vs author (year). Also dont capitalize thermodynamic integration after defining the acronym.
