### 2023-11-03 #5886 @demonico85
> I am also wondering if you could comment on your choice of distribution / installation method. Why did you decide against contributing your work to LAMMPS directly? I think that would have many benefits, including a review from the LAMMPS maintainers, reducing the chance of integration problems (for the current or a future version), as well as a simpler installation process for users. Downloading a third-party repo and copying source files into the LAMMPS distribution certainly works, but isn't the greatest user experience, and might be especially challenging when trying to use the package on HPC resources.

We agree with the reviewer's comment about the less convenient user experience in having a separate package with respect to an addition to the main LAMMPS repository. However, we did not decide against the inclusion in LAMMPS directly. According to LAMMPS developers (https://docs.lammps.org/Modify_contribute.html), external packages are equally welcomed to increase the potentiality of LAMMPS itself. Therefore, we decided, for the moment, to stick with this form of contribution rather than submitting the package to the LAMMPS developer, creating a package with the core code needed to run the cleaving model. Future extensions of this core model with different type of interactions (e.g., those used in water model) will make the package general enough for us to consider to request to LAMMPS developer to be included in the main distribution
