### 2023-11-03 #5886 @demonico85
> Another point Re: "Contribution and authorship: Has the submitting author (@demonico85) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?". It's easy to verify through the git log that @demonico85 and @lorenzo-rovigatti (first and last author) have contributed significantly to the software. The git log does not allow to verify Ruslan Davidchacks contribution. Can you comment on including them as an author?

Most of the code we are presenting here, which was developed in this form mainly by myself and [@lorenzo-rovigatti](https://github.com/lorenzo-rovigatti) is based on a in-house Fortran code written by the third author: Ruslan Davidchack. For this reason, despite the obvious differences (Fortran VS c++, in-house MD code with cleaving VS package for LAMMPS) the contribution of the third author to the code is substantial, even if he is not listed as author of the repository itself. In our opinion, and for the sake of fairness, we feel that he needs to be included as an author of the repository
