### 2023-11-03 #5886 @demonico85
> Here's a few more things I noticed:

> * **utils:** You have a large number of files in the `utils/` folder, which make up a non-trivial part of your source code. Unfortunately, most of these are not documented. I see that 3 of them are described in the last step of the examples, but one of them (`calcSFE.m`) isn't even present in my checkout. This leaves 40 files without any description. Do you consider these files part of this submission? In this case, I think these would need a description in the documentation, and ideally tests.

The `utils/` folder contains a miscellanea of programs that can help in working with the cleaving model but they are not part of the model itself (e.g., bash files to run the LAMMPS simulations on a queue system). We included them to avoid the users to write their own short scripts (e.g., to collect the work in step3) if they want, but since this is not part of the cleaving model we did not include any explanation in the documentation. The reason is that if a user finds it easier to write a bash script using awk to extract the work in step3, rather than adapting the fortran code provided in the `/utils` folder, that's perfectly reasonable.

As per the calcSFE.m subroutine, we corrected the description of the guide related to the examples. That particular subroutine is located in the folder `/examples/LJ_SL/utils` (for the example on solid-liquid interface) and in the folder `/examples/LJ_SV/utils` (for the example on solid-vacuum interface). We wanted to make the example as self-contained as possible, so each of them has their own post-processing subroutines (adapted for that particular example).

However, we agree that it maybe confusing to have a bunch of code without any explanation at all. We included in the guide a  list of the subroutines in `utils/` with a single explanation line to help the user to navigate through them.
