### 2023-11-03 567a761
Cleaned utils
