### 2023-11-05 010d8c6
Fix two citations, as per https://github.com/openjournals/joss-reviews/issues/5886
