### 2023-11-07 e062599
Fixed capital case TI
