+++ joss/paper.md
@@ -49,7 +49,7 @@ Direct determination of the IFE from the reversible work of creating an interfac
 
 ![The thermodynamic path used to compute the interfacial free energy between phases $\alpha$ and $\beta$ with the cleaving method\label{fig:cleaving}](Fig/joss.png){ width=30% }
 
-A sketch of the cleaving path to obtain the IFE between two generic phases $\alpha$ and $\beta$ is shown in \autoref{fig:cleaving}. The reversible work in each step can be determined by different free energy computation methods, but those based on Thermodynamic Integration are the most straightfoward and provide accurate and direct results. With TI methods the free energy difference between two thermodynamic states connected by a transformation path is calculated by integrating the ensemble average of some configuration dependent function (*e.g.*, the potential energy) with respect to a parameter defining the path from the initial to the final thermodynamic state.
+A sketch of the cleaving path to obtain the IFE between two generic phases $\alpha$ and $\beta$ is shown in \autoref{fig:cleaving}. The reversible work in each step can be determined by different free energy computation methods, but those based on thermodynamic integration are the most straightfoward and provide accurate and direct results. With TI methods the free energy difference between two thermodynamic states connected by a transformation path is calculated by integrating the ensemble average of some configuration dependent function (*e.g.*, the potential energy) with respect to a parameter defining the path from the initial to the final thermodynamic state.
 
 
 To date, the cleaving model, which owes its name to the fact that each bulk phase is "cleaved", *i.e.*, cut at a certain position to create the new interface, has been used to calculate the solid-liquid IFE in hard and soft spheres [@Davidchack00prl;@Davidchack05prl], Lennard-Jones
@@ -59,7 +59,7 @@ To date, the cleaving method, which owes its name to the fact that each bulk pha
 
 # Functionality
 
-CLEAVING is a LAMMPS package for the calculation of IFE using Thermodynamic Integration. The package includes several new additional functions written in the style suited to be directly patched into LAMMPS, comprising interactions between atoms needed to run the MD simulation with the cleaving model and some auxiliary functionalities (in LAMMPS terms, new "fix" and "compute") needed to perform the operations required to carry out the different steps of the calculation. 
+CLEAVING is a LAMMPS package for the calculation of IFE using thermodynamic integration. The package includes several new additional functions written in the style suited to be directly patched into LAMMPS, comprising interactions between atoms needed to run the MD simulation with the cleaving model and some auxiliary functionalities (in LAMMPS terms, new "fix" and "compute") needed to perform the operations required to carry out the different steps of the calculation. 
 
 The version of the code we present here is hosted on GitHub, where a detailed instruction manual is complemented by some step-by-step examples reproducing some of the published results on the topic. 
 The package includes new definitions of pair potentials not included in LAMMPS, as the Broughton-Gilmer modified Lennard-Jones potential [@Broughton1983] and modifications of some of the already existing pair_styles in LAMMPS. In particular, we included a modified version of the `lj/cut` and `coul/dsf` potentials. These modifications allow to execute the third step of the cleaving method, where the interactions between the phases are switched on and off. The difficulty in this case is to keep track of the interactions between atoms involved in the switching, since the work performed during this particular step depends on them. Moreover, these new `pair_style`s support different scaling (*i.e.*, different powers of the coupling parameter $\lambda$, see *e.g.*, [@DiPasquale2022cleaving]).
