### 2023-11-07 #5886 @demonico85
> The work implements an accessible cleaving method in LAMMPS which will certainly be of interest for the readership of JOSS. Overall the work is done carefully, the documentation is clearly written and figures are informative. Adding to what has been put forth ptmerz already, here are some additional comments:
>
> **Functionality** The main point that I would like discussed in more depth is the range of applicability of the software. What I mean is that in the documentation and paper we learn about prior uses of the method including hard and soft-spheres, LJ, TIP4P, EAM and others and that the adoption for different forcefields requires a more convenient simulation interface. Which is the motivation for this work. To me this suggested that the software can already cleave TIP4P, EAM and others, but I am not sure if this is the case since in TIP4P, Handel and Davichack et al. developed an orientational cleaving potential for step 2, and in the case of EAM, Liu and coworkers added a soft repuslive part to the potential in order to compensate their calculation at non-coexistence, again to achieve crystal layers in step 2. Since I do not see this implemented, could you clarify in the documentation and paper what the current implementation is capable of? Say I want to repeat Handel et al.s work, can I:
>
> 1. use existing pair_styles and fixes (it would be great to have an example in this case)
> 2. or do I need edit wellPforce to include orientation (add some documentation how to modify fixes for custom force-fields)
> 3. or can I not use this package for this purpose yet (then state this more clearly)
>
> If the latter is the case and its not possible to do step 2 for arbitrary systems, is it still possible to do step 1,3 and 4? So the solid-vacuum TIP4P interface. TIP4P can be made with lj/cut/coul and fix rigid and if that all works with your implementation it would be awesome to add an example of this more complicated scenario.

The reviewer is correct in pointing out the fact that the package does not include the capabilities to run the work of Handel et al.

1. The fix wellPforce does not include orientation

2. The pair styles currently considered in the repository do not include the TIP4P for water.

We have added the following sentences to the paper

`Although the package in its current form does not include yet all the different interactions explored in the literature, it allows to reproduce the results in [@Davidchack03direct;@DiPasquale2022cleaving;@DiPasquale2020shuttleworth] and gives all the machinery needed to extend the calculations to more complex interaction models. The package we are presenting makes the addition of new interactions type a straightforward extension of the code already given here.`

to make clear what the package can achieve at the moment, after listing the results obtained with the cleaving model. In our opinion, it is still an important piece of information showing the capabilities and possible range of application of the cleaving model, to advocate for the importance of the current repository.

Having said that, the important concept for our repository is that it represents the core functionalities on top of which more complicated models can be built on. Step 1, 2, 4 can be performed for arbitrary interactions, because what is needed for the cleaving are the external potentials (walls or wells).  Instead, for step3 some modifications of the subroutines calculating the interactions are needed. The machinery for step3 to switch on/off the interactions is already in the repository (i.e., the move/dupl, duplicate), but a new interaction style not already included in the repository would need to be modified to include the scaling of the interactions with lambda (as done e.g., for pair_lj_BGcleav.cpp). However, these modifications just require adapting the already existing subroutines to the new interactions one wishes to consider. The repository as it stands at the moment allows either to run directly  the cleaving model with LAMMPS (with a limited selection of interactions, LJ, Broughton and Gilmer LJ, Fennel potential (dsf)) or to provide a firm basis for extension to more complicated/exotic interactions.

As a side note, we are already working in extending the cleaving repository with TIP4P interactions. As we are still testing the results, we put them in a separate branch (graphene) of the repository. They will become part of the repository as soon as they are ready.

We want to highlight here that this last remark is important to see the importance of what we are presenting at the moment. Extensions of the repository will include more different interactions, which will be added in the future, but the core functionalities for the cleaving model (on top of which everything else can be built) are the one we are presenting with this submission.

>
> Minor points:
>
> * In evaluating step3 of examples/lj_SV using a.out I got ERROR file zdir.dat not found. I copied zdir.dat into /out/ and it worked. All other steps in the example worked without issue!

Corrected the documentation, the file zdir.dat must be copied in the dir from where a.out is called

> * Does it have to be Matlab to calculate the integrals? Something non-proprietary would be preferable, like the calcintegrals.f90 in utils.

No, Matlab is not required. It is only a numerical calculation of an integral, for the sake of the cleaving model it does not matter the way it is performed. We included the matlab file as it was the one we used to derive the results published in the paper on which the example are based [Di Pasquale & Davidchack 2020]

> * Speaking of utils, there is no documentation for these scripts but some sound very useful for users.

We now included in the documentation a brief description of each subroutine in the utils folder. As we discussed in another reply, since these subroutines are not part of the cleaving model (they are basically pre/post processing) we did not want to "force" them on the user. For example, if Matlab is not an option, any python/fortran/excel/... subroutine for the calculation of the integral is ok

> * The citation style on line 68-73 is not uniform (author, year) vs author (year). Also dont capitalize thermodynamic integration after defining the acronym.

fixed the citation style
