### 2023-11-09 57fc194
Add statement of need in the documentation (as requested in the [JOSS review](https://github.com/openjournals/joss-reviews/issues/5886))
