### 2023-11-09 1282266
Add community guidelines to the docs
This commit is linked to the JOSS review (https://github.com/openjournals/joss-reviews/issues/5886)
