+++ docs/source/index.md
@@ -28,6 +28,7 @@ To date, the cleaving method, which owes its name to the fact that each bulk pha
    lammps.md
    example.md
    utils.md
+   contributing.md
 ```
 
 ## Bibliography
