### 2023-11-11 9fe821a
Added regression test (as requested in the [JOSS review](openjournals/joss-reviews#5886) )
