### 2023-11-11 9ea96b9
Change the name of the testing scripts and update the documentation
