### 2023-11-11 0a684b7
Fix the testing script
