### 2023-11-12 b2e7773
Change the full testing script to make it run longer
