### 2023-11-12 89a3b81
Fix the computation of the averages of the full test
