+++ test/cleav_int.sh
@@ -1,0 +1,465 @@
+#! /bin/bash -l
+#
+
+function checkcommands () {
+
+if ! command -v $1 &> /dev/null
+then
+    echo "ERROR: $1 could not be found"
+    exit
+fi
+
+}
+
+
+# Parameters to be changed
+
+lmp="lmp_mpi"
+
+trap "exit 15" TERM
+export TOP_PID=$$
+
+checkcommands "mpirun"
+checkcommands "gfortran"
+
+# ------------ Init the systems ------------------------------------------------
+
+
+
+runsim=1
+nproc=4
+re='^[0-9]+([.][0-9]+)?$'
+continue=0
+d=$(date +%Y-%m-%d)
+
+echo "Starting of the simulation" 
+date +"%c"
+echo
+
+# Read Inputs
+
+while getopts l:n:h option
+do
+ case "${option}"
+ in
+ l) lmp=${OPTARG}
+    ;;
+ n) nproc=${OPTARG}
+    if ! [[ $nproc =~ $re ]] ; then
+       echo "ERROR: Input for -n not an integer"
+       echo "Exiting ..." 
+       exit 1
+    fi
+    ;;
+ h) echo "cleav_regression.sh -l <path lammps executable> OPTIONAL -n <nproc, default=4>"
+    ;;
+ \?)
+    echo "ERROR: Invalid option: -$OPTARG" 
+    echo "Exiting ..."
+    exit 4
+    ;;
+ esac
+done
+
+lmp=$(realpath $lmp)
+checkcommands $lmp
+
+if ! [[ $nproc =~ $re ]] ; then
+   echo "ERROR: Need to include number of processes (option -n)"
+   echo "Exiting ..."
+   exit 1
+fi
+
+echo "Number of MPI processes: $nproc"
+
+
+# Check if all the commands are available
+
+
+
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+#*******************************************************************************
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+# Check if all the directories are in the correct place
+################################################################################
+
+
+
+echo "I am working with executable $lmp"
+
+
+if [ ! -d $exampledir ]; then
+    echo "ERROR: I cannot find $exampledir and I cannot run the test"
+    echo "Exiting..."
+    echo -1
+fi
+
+
+now=$(date +"%m_%d_%Y_%s")
+log=$(echo $now.log)
+
+#################################################################################
+
+echo "Starting regression test 1"
+
+if [ ! -d iRT1 ];
+  then
+   echo "I cannot find dir R1 to run the test"
+   echo "Check your directory"
+   echo "Exiting..."
+   exit
+fi
+
+rm -rf test_RT1
+cp -r iRT1 test_RT1
+
+echo "Entering working dir..."
+cd ./test_RT1
+echo $(pwd)
+echo
+
+CWD=$(pwd)
+
+
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+# STEP 1
+
+step="step1"
+
+echo $step
+echo
[…]
