+++ test/iRT2/step3/step3.in
@@ -1,0 +1,167 @@
+###%***************************************************************************%
+###%                                                                          *%                                                                                                                   
+###%    Script to run Step 1 of the Cleaving Method with Walls                *%
+###%                                                                          *%
+###%***************************************************************************%
+###%                                                                          *%
+###%    Based on:                                                             *%
+###%    R. Davidchack, B. Laird, J. Chem. Phys., 118:7651-7657 (2003)         *%
+###%                                                                          *%
+###%   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~         *%
+###%                                                                          *%
+###%    Created by:                                                           *%
+###%    Nicodemo Di Pasquale                                                  *%
+###%    Department of Mathematics                                             *%
+###%    University of Leicester                                               *%
+###%    24 March 2018                                                         *%
+###%                                                                          *%
+###%***************************************************************************%
+
+
+units		lj
+atom_style	atomic
+
+boundary p p p
+read_data inputStep3.lmp
+
+mass * 1.0
+
+
+group crys type 1 4 5 6
+group liquid type  2 3 7 8 
+
+group real type 1 2 3 4
+group dupl type 5 6 7 8
+
+group realliq type 2 3
+group realcry type 1 4
+
+# ---------------------------- Define variables --------------------------------
+
+variable clwall1 equal 8.099840380596101
+variable clwall2 equal 24.299521141788304 
+variable backloop equal 0
+variable  sigma   equal  1.0
+variable  eps     equal  1.0
+variable  delta   equal  0.25
+variable  siglj   equal  1.0
+variable  epslj   equal  1.0
+variable cnt equal 1
+variable lam file lambda.dat 
+variable lambda equal ${lam}
+variable delam equal 0.005
+variable zwf equal 0.62
+variable  crw     equal  2.0^(1.0/6.0)   
+variable  rhoc    equal  0.945
+variable  cut1    equal  2.3
+variable  cut2    equal  2.5
+variable  eqnts   equal  0
+variable  nts     equal  1000
+variable  ts      equal  0.005
+variable thermoSteps equal 5000
+variable dumpfreq equal 20000
+variable Nevery equal 1
+variable Nrepeat equal 1 
+variable Nfreq equal 1000
+variable  nloop   equal  2000
+variable Tsyst equal 0.617
+variable  Tdump   equal  0.0748
+variable restartfreq equal 200000 
+
+# ------------------------- Derivate variables ---------------------------------
+#
+
+variable minlam equal 1-${lambda}
+variable  rw    equal  ${sigma}*${crw}
+variable  cutoff1 equal  ${siglj}*${cut1} 
+variable  cutoff2 equal  ${siglj}*${cut2}
+
+# ---------------------------- Neighbours Exclusion ----------------------------
+
[…]

[…]
