### 2023-11-13 817bdb8
Add a new interaction test
The new test can be used as a "regression test" as discussed in the JOSS review (https://github.com/openjournals/joss-reviews/issues/5886)
