+++ _sources/contributing.md.txt
@@ -1,0 +1,5 @@
+# Contributing
+
+If you want to add functionalities (or fix issues) to CLEAVING, you are more than welcome to do so. Feel free to fork the repo and open a pull request: we will do our best to review it. Note that if we accept your changes you will be asked to update the documentation to reflect what you have done.
+
+If you have any questions, find a bug or an issue or want to propose a new feature, please do so by creating a [new issue](https://github.com/demonico85/cleaving/issues/new/choose). Pick the issue template that makes the most sense and edit it to add as much detail as possible.
