### 2023-11-13 dcc70ed
Conclude the set up of the interaction test
