+++ docs/source/install.md
@@ -16,16 +16,16 @@ The list of known issues can be browsed online [here](https://github.com/demonic
 
 ## Testing
 
-The installation can be tested by heading over to `/test`. We provide two different tests: the first one checks that the installation (and therefore the main LAMMPS cleaving commands) works by performing one-time-step simulations, while the second one runs slightly longer simulations to perform ``interaction'' checks for each cleaving step for the solid-liquid and solid-vacuum Broughton and Gilmer LJ potential.
+The installation can be tested by heading over to `/test`. We provide two different tests: the first one checks that the installation (and therefore the main LAMMPS cleaving commands) works by performing one-time-step simulations, while the second one runs slightly longer simulations to perform "interaction" checks for each cleaving step for the solid-liquid and solid-vacuum Broughton and Gilmer LJ potential.
 
-The tests can be run by executing `./cleav_run.sh` (for the installation check) or `./cleav_int.sh` (for the interaction test) following these instructions:
+The tests can be run by executing `./cleav_run.sh` (for the installation check) or `./cleav_int.sh` (for the interaction test):
 
 1. LAMMPS should be patched with the cleaving files and compiled as detailed [above](#compilation).
 2. Run the chosen script with a `-l LAMMPS_BINARY` argument, where `LAMMPS_BINARY`, which defaults to `lmp_mpi`, is the path of the LAMMPS executable to be tested.
 3. An optional argument `-n` can be given to select the number of the MPI processes to be used, which defaults to 4.
-4. Both tests should take a few minutes on 4 cores. 
+4. Either test should run in less than a minute on 4 cores. 
 
-Please, note that the results obtained in either case are *not* going to be physically correct or meaningful (for which we refer to the [examples](example.md)). For `cleav_run.sh`, the script only tests if the pairs/fix/compute of the cleaving package run without errors. By contrast, `cleav_int.sh` tests if the cleaving code gives consistent results, and can be used to check if any modifications affects the behaviour in an unwanted way. The final output of the latter test (which is printed at the end of the procedure, but also stored in the `test_iRT1/results.dat` and `test_iRT2/results.dat` files) should be:
+Please, note that the results obtained in either case are *not* going to be physically correct or meaningful (for which we refer to the [examples](example.md)). For `cleav_run.sh`, the script only tests if the pairs/fix/compute of the cleaving package run without errors. By contrast, `cleav_int.sh` tests if the cleaving code gives consistent results, and should be used to check if any modifications affects the behaviour in an unwanted way. The final output (which is printed at the end of the procedure, but also stored in the `test_iRT1/results.dat` and `test_iRT2/results.dat` files) should be:
 
     #####################################
     Results from iRT1 (solid-vacuum LJ):
