### 2023-11-13 #5886 @lorenzo-rovigatti
We have added a new testing script that runs short cleaving steps and output the resulting work. The numbers can be used to check for any regression bugs, as requested by @ptmerz . We have update the documentation to [reflect this change](https://demonico85.github.io/cleaving/install.html#testing).
We have also added contributing guidelines (both to the readme and to the docs), as well as added a background/statement of need part to the [docs](https://demonico85.github.io/cleaving/#background). If I'm not mistaken these were the last things that we were asked to discuss/change!
