### 2023-11-30 1dbac01
Fixed bug on the calculation of the position m in the pvector
