+++ src/pppm_tip4p_cleav.cpp
@@ -1,0 +1,699 @@
+// clang-format off
+/* ----------------------------------------------------------------------
+   LAMMPS - Large-scale Atomic/Molecular Massively Parallel Simulator
+   https://www.lammps.org/, Sandia National Laboratories
+   LAMMPS development team: developers@lammps.org
+
+   Copyright (2003) Sandia Corporation.  Under the terms of Contract
+   DE-AC04-94AL85000 with Sandia Corporation, the U.S. Government retains
+   certain rights in this software.  This software is distributed under
+   the GNU General Public License.
+
+   See the README file in the top-level LAMMPS directory.
+------------------------------------------------------------------------- */
+
+/* ----------------------------------------------------------------------
+   Contributing authors: Amalie Frischknecht and Ahmed Ismail (SNL)
+------------------------------------------------------------------------- */
+
+#include "pppm_tip4p_cleav.h"
+
+#include "atom.h"
+#include "domain.h"
+#include "force.h"
+#include "error.h"
+#include "math_const.h"
+
+#include <cmath>
+
+using namespace LAMMPS_NS;
+using namespace MathConst;
+
+#define OFFSET 16384
+
+#ifdef FFT_SINGLE
+#define ZEROF 0.0f
+#define ONEF  1.0f
+#else
+#define ZEROF 0.0
+#define ONEF  1.0
+#endif
+
+/* ---------------------------------------------------------------------- */
+
+PPPMTIP4PCleav::PPPMTIP4PCleav(LAMMPS *lmp) : PPPM(lmp)
+{
+  triclinic_support = 1;
+  tip4pflag = 1;
+}
+
+/* ---------------------------------------------------------------------- */
+
+void PPPMTIP4PCleav::settings(int narg, char **arg)
+{
+  if (narg < 2) error->all(FLERR,"Illegal kspace_style {} command", force->kspace_style);
+
+  accuracy_relative = fabs(utils::numeric(FLERR,arg[0],false,lmp));
+  if (accuracy_relative > 1.0)
+    error->all(FLERR, "Invalid relative accuracy {:g} for kspace_style {}",
+               accuracy_relative, force->kspace_style);
+               
+  lambda = utils::numeric(FLERR,arg[1],false,lmp);
+  
+}
+
+/* ---------------------------------------------------------------------- */
+
+
+void PPPMTIP4PCleav::init()
+{
+  // TIP4P PPPM requires newton on, b/c it computes forces on ghost atoms
+
+  if (force->newton == 0)
+    error->all(FLERR,"Kspace style pppm/tip4p requires newton on");
+
+  PPPM::init();
+}
+
+/* ----------------------------------------------------------------------
+   find center grid pt for each of my particles
+   check that full stencil for the particle will fit in my 3d brick
+   store central grid pt indices in part2grid array
+------------------------------------------------------------------------- */
+
+void PPPMTIP4PCleav::particle_map()
+{
+  int nx,ny,nz,iH1,iH2;
[…]
