### 2023-12-04 d8f95ac
Fixed pair_lj_cut_tip4p_short with correct scaling lambda
