+++ src/pair_lj_cut_tip4p_short_cleav.cpp
@@ -19,6 +19,8 @@
 
 #include "pair_lj_cut_tip4p_short_cleav.h"
 
+#include "compute_chunk_atom.h"
+#include "compute_com_chunk.h"
 #include "angle.h"
 #include "atom.h"
 #include "bond.h"
@@ -29,6 +31,7 @@
 #include "neigh_list.h"
 #include "memory.h"
 #include "error.h"
+#include "modify.h"
 
 #include <cmath>
 #include <cstring>
@@ -45,7 +48,7 @@ using namespace LAMMPS_NS;
 
 /* ---------------------------------------------------------------------- */
 
-PairLJCutTIP4PLongCleav::PairLJCutTIP4PLongCleav(LAMMPS *lmp) :
+PairLJCutTIP4PShortCleav::PairLJCutTIP4PShortCleav(LAMMPS *lmp) :
   PairLJCutCoulLong(lmp)
 {
   tip4pflag = 1;
@@ -54,8 +57,6 @@ PairLJCutTIP4PLongCleav::PairLJCutTIP4PLongCleav(LAMMPS *lmp) :
   single_enable = 0;
   respa_enable = 0;
   writedata = 1;
-  pallocation = 0;
-
 
   nmax = 0;
   hneigh = nullptr;
@@ -67,13 +68,16 @@ PairLJCutTIP4PLongCleav::PairLJCutTIP4PLongCleav(LAMMPS *lmp) :
   no_virial_fdotr_compute = 1;
   
   dubtypes=ntypes*2;
-  if(n > 0){
-    nextra = n*n+1; // the zero-th position is not considered
+
+
+  pallocation = 0;
+  natoms=natoms+1;
+  if(ntypes > 0){
+    nextra = dubtypes*dubtypes+1; // the zero-th position is not considered
     pvector = new double[nextra];
-    pallocation=1;
+    pallocation = 1;
     }
-    
-    
+
    xprd = domain->xprd;
    yprd = domain->yprd;
    zprd = domain->zprd;
@@ -81,15 +85,17 @@ PairLJCutTIP4PLongCleav::PairLJCutTIP4PLongCleav(LAMMPS *lmp) :
    yz = domain->yz;
    xz = domain->xz;
 
+  nchunk = 0;  
 }
 
 /* ---------------------------------------------------------------------- */
 
-PairLJCutTIP4PLongCleav::~PairLJCutTIP4PLongCleav()
+PairLJCutTIP4PShortCleav::~PairLJCutTIP4PShortCleav()
 {
   memory->destroy(hneigh);
   memory->destroy(newsite);
   
+  if (allocated) {  
     memory->destroy(lam);
     memory->destroy(powlambda);
     memory->destroy(powDlambda);
@@ -98,18 +104,54 @@ PairLJCutTIP4PLongCleav::~PairLJCutTIP4PLongCleav()
     memory->destroy(powlambdaC);
     memory->destroy(powDlambdaC);
     memory->destroy(giflag);
-    
-  pallocation = 0;
+    }
+  
+}
+
+/* ----------------------------------------------------------------------
+   allocate all arrays
+------------------------------------------------------------------------- */
+
+void PairLJCutTIP4PShortCleav::allocate()
+{
+  allocated = 1;
+  int n = atom->ntypes;
+
+  memory->create(setflag,n+1,n+1,"pair:setflag");
+  for (int i = 1; i <= n; i++)
+    for (int j = i; j <= n; j++)
+      setflag[i][j] = 0;
+
+  memory->create(cutsq,n+1,n+1,"pair:cutsq");
+
+  memory->create(cut_lj,n+1,n+1,"pair:cut_lj");
[…]
