### 2023-12-05 5aefa55
pair potential containing only the short range part of the full coulombic
Need to modify pair_lj_tip4p_short_cleav
