### 2023-12-06 e4d708f
Corrected function for short range coulombic with cleaving
