+++ src/pair_lj_cut_tip4p_long_cleav.h
@@ -13,21 +13,21 @@
 
 #ifdef PAIR_CLASS
 // clang-format off
-PairStyle(lj/cut/tip4p/short/cleav,PairLJCutTIP4PShortCleav);
+PairStyle(lj/cut/tip4p/long/cleav,PairLJCutTIP4PLongCleav);
 // clang-format on
 #else
 
-#ifndef LMP_PAIR_LJ_CUT_TIP4P_SHORT_CLEAV_H
-#define LMP_PAIR_LJ_CUT_TIP4P_SHORT_CLEAV_H
+#ifndef LMP_PAIR_LJ_CUT_TIP4P_LONG_CLEAV_H
+#define LMP_PAIR_LJ_CUT_TIP4P_LONG_CLEAV_H
 
 #include "pair_lj_cut_coul_long.h"
 
 namespace LAMMPS_NS {
 
-class PairLJCutTIP4PShortCleav : public PairLJCutCoulLong {
+class PairLJCutTIP4PLongCleav : public PairLJCutCoulLong {
  public:
-  PairLJCutTIP4PShortCleav(class LAMMPS *);
-  ~PairLJCutTIP4PShortCleav() override;
+  PairLJCutTIP4PLongCleav(class LAMMPS *);
+  ~PairLJCutTIP4PLongCleav() override;
   void compute(int, int) override;
   void settings(int, char **) override;
   void init_style() override;
@@ -50,11 +50,12 @@ class PairLJCutTIP4PShortCleav : public PairLJCutCoulLong {
   char *idchunk, *idcom;  
   class ComputeChunkAtom *cchunk;
   class ComputeCOMChunk *ccom;
-      
+ 
+ 
+ 
   int typeH, typeO;    // atom types of TIP4P water H and O atoms
   int typeA, typeB;    // angle and bond types of TIP4P water
   double alpha;        // geometric constraint parameter for TIP4P
-
 
   int nmax;            // info on off-oxygen charge sites
   int **hneigh;        // 0,1 = indices of 2 H associated with O
@@ -62,6 +63,8 @@ class PairLJCutTIP4PShortCleav : public PairLJCutCoulLong {
   double **newsite;    // locations of charge sites
 
   void compute_newsite(double *, double *, double *, double *);
+  
+  
   void global_boundary();
   int find_scaling(int,int,int,int,double *);
   void allocate() override;  
