+++ src/pair_lj_cut_tip4p_long_cleav.cpp
@@ -17,7 +17,7 @@
    simpler force assignment added by Rolf Isele-Holder (Aachen University)
 ------------------------------------------------------------------------- */
 
-#include "pair_lj_cut_tip4p_short_cleav.h"
+#include "pair_lj_cut_tip4p_long_cleav.h"
 
 #include "compute_chunk_atom.h"
 #include "compute_com_chunk.h"
@@ -48,7 +48,7 @@ using namespace LAMMPS_NS;
 
 /* ---------------------------------------------------------------------- */
 
-PairLJCutTIP4PShortCleav::PairLJCutTIP4PShortCleav(LAMMPS *lmp) :
+PairLJCutTIP4PLongCleav::PairLJCutTIP4PLongCleav(LAMMPS *lmp) :
   PairLJCutCoulLong(lmp)
 {
   tip4pflag = 1;
@@ -67,11 +67,20 @@ PairLJCutTIP4PShortCleav::PairLJCutTIP4PShortCleav(LAMMPS *lmp) :
 
   no_virial_fdotr_compute = 1;
   
-  dubtypes=ntypes*2;
+  ntypes=atom->ntypes;
+  natoms=atom->natoms; //array does start from zero
 
+
+
+  if(natoms == 0){
+      error->all(FLERR,"Pair style must be declared after box generation");
+  }
+  
 
   pallocation = 0;
   natoms=natoms+1;
+  dubtypes=ntypes*2;
+      
   if(ntypes > 0){
     nextra = dubtypes*dubtypes+1; // the zero-th position is not considered
     pvector = new double[nextra];
@@ -86,14 +95,16 @@ PairLJCutTIP4PShortCleav::PairLJCutTIP4PShortCleav(LAMMPS *lmp) :
    xz = domain->xz;
 
   nchunk = 0;  
+  
 }
 
 /* ---------------------------------------------------------------------- */
 
-PairLJCutTIP4PShortCleav::~PairLJCutTIP4PShortCleav()
+PairLJCutTIP4PLongCleav::~PairLJCutTIP4PLongCleav()
 {
   memory->destroy(hneigh);
   memory->destroy(newsite);
+  
   
   if (allocated) {  
     memory->destroy(lam);
@@ -105,14 +116,15 @@ PairLJCutTIP4PShortCleav::~PairLJCutTIP4PShortCleav()
     memory->destroy(powDlambdaC);
     memory->destroy(giflag);
     }
-  
+    
 }
+
 
 /* ----------------------------------------------------------------------
    allocate all arrays
 ------------------------------------------------------------------------- */
 
-void PairLJCutTIP4PShortCleav::allocate()
+void PairLJCutTIP4PLongCleav::allocate()
 {
   allocated = 1;
   int n = atom->ntypes;
@@ -140,13 +152,23 @@ void PairLJCutTIP4PShortCleav::allocate()
   memory->create(powlambdaC,n+1,n+1,"pair:powlambdaC");
   memory->create(powDlambdaC,n+1,n+1,"pair:powDlambdaC");
   memory->create(giflag,natoms,"pair:giflag");  
+  memory->create(gbox,natoms,"pair:gbox");  
   
-  
+
+  ntypes=n;
+  dubtypes=n*2;
+  if(!pallocation){
+    nextra = dubtypes*dubtypes+1; // the zero-th position is not considered
+    pvector = new double[nextra];
+    pallocation = 1;
+    }
+
 }
+
 
 /* ---------------------------------------------------------------------- */
 
-void PairLJCutTIP4PShortCleav::compute(int eflag, int vflag)
+void PairLJCutTIP4PLongCleav::compute(int eflag, int vflag)
 {
[…]
