### 2023-12-09 52b1aa4
Removed unused functions declarations from pppm_tip4p_dummy.cpp file
