+++ src/pppm_tip4p_dummy.cpp
@@ -67,45 +67,7 @@ void PPPMTIP4Pdummy::init()
 
 void PPPMTIP4Pdummy::particle_map()
 {
-  int nx,ny,nz,iH1,iH2;
-  double *xi,xM[3];
 
-  int *type = atom->type;
-  double **x = atom->x;
-  int nlocal = atom->nlocal;
-
-  if (!std::isfinite(boxlo[0]) || !std::isfinite(boxlo[1]) || !std::isfinite(boxlo[2]))
-    error->one(FLERR,"Non-numeric box dimensions - simulation unstable");
-
-  int flag = 0;
-  for (int i = 0; i < nlocal; i++) {
-    if (type[i] == typeO) {
-      find_M(i,iH1,iH2,xM);
-      xi = xM;
-    } else xi = x[i];
-
-    // (nx,ny,nz) = global coords of grid pt to "lower left" of charge
-    // current particle coord can be outside global and local box
-    // add/subtract OFFSET to avoid int(-0.75) = 0 when want it to be -1
-
-    nx = static_cast<int> ((xi[0]-boxlo[0])*delxinv+shift) - OFFSET;
-    ny = static_cast<int> ((xi[1]-boxlo[1])*delyinv+shift) - OFFSET;
-    nz = static_cast<int> ((xi[2]-boxlo[2])*delzinv+shift) - OFFSET;
-
-    part2grid[i][0] = nx;
-    part2grid[i][1] = ny;
-    part2grid[i][2] = nz;
-
-    // check that entire stencil around nx,ny,nz will fit in my 3d brick
-
-    if (nx+nlower < nxlo_out || nx+nupper > nxhi_out ||
-        ny+nlower < nylo_out || ny+nupper > nyhi_out ||
-        nz+nlower < nzlo_out || nz+nupper > nzhi_out) flag++;
-  }
-
-  int flag_all;
-  MPI_Allreduce(&flag,&flag_all,1,MPI_INT,MPI_SUM,world);
-  if (flag_all) error->all(FLERR,"Out of range atoms - cannot compute PPPM");
 }
 
 /* ----------------------------------------------------------------------
@@ -117,54 +79,7 @@ void PPPMTIP4Pdummy::particle_map()
 
 void PPPMTIP4Pdummy::make_rho()
 {
-  int i,l,m,n,nx,ny,nz,mx,my,mz,iH1,iH2;
-  FFT_SCALAR dx,dy,dz,x0,y0,z0;
-  double *xi,xM[3];
 
-  // clear 3d density array
-
-  FFT_SCALAR *vec = &density_brick[nzlo_out][nylo_out][nxlo_out];
-  for (i = 0; i < ngrid; i++) vec[i] = ZEROF;
-
-  // loop over my charges, add their contribution to nearby grid points
-  // (nx,ny,nz) = global coords of grid pt to "lower left" of charge
-  // (dx,dy,dz) = distance to "lower left" grid pt
-  // (mx,my,mz) = global coords of moving stencil pt
-
-  int *type = atom->type;
-  double *q = atom->q;
-  double **x = atom->x;
-  int nlocal = atom->nlocal;
-
-  for (int i = 0; i < nlocal; i++) {
-    if (type[i] == typeO) {
-      find_M(i,iH1,iH2,xM);
-      xi = xM;
-    } else xi = x[i];
-
-    nx = part2grid[i][0];
-    ny = part2grid[i][1];
-    nz = part2grid[i][2];
-    dx = nx+shiftone - (xi[0]-boxlo[0])*delxinv;
-    dy = ny+shiftone - (xi[1]-boxlo[1])*delyinv;
[…]
