### 2023-12-23 830e950
adjusted allocation with 2*nytpes
