### 2024-01-15 #5886 @ptmerz
I first want to thank the authors for the improvements they have implemented in response to the reviewer comments. I have finished the outstanding review tasks to complete the checklist, and I think the work presented here is relevant to the field and the software quality is good, making it overall well suited for JOSS.
