### 2024-01-25 #5886 @lorenzo-rovigatti
Thank you @mbarzegary !
I merged the PR and uploaded the 1.0 version to Zenodo (DOI: 10.5281/zenodo.10567702).
