### 2024-01-25 #5886 @mbarzegary
@lorenzo-rovigatti can you please make the license of the archive similar to the license of the software repository?
