### 2024-01-25 #5886 @lorenzo-rovigatti
Good catch, sorry about that! I changed the license to GPL 3.0, which is the same we use for the code in the repo.
