### 2024-01-26 #5886 @lorenzo-rovigatti
I have amended the author list order. As for the version, I see v1.0 everywhere on the zenodo page. What did I miss?
