### 2024-01-29 #5886 @Kevin-Mattheus-Moerman
@demonico85 as AEiC I will now help to process this work for acceptance in JOSS. I have checked your repository, the archive link, this review, and the paper. Most seems in order, I only have the below minor points that require your attention.

- [ ] `characterisation` is British English while `generalizable` is American English. Consider making the language choice consistent.
- [ ] `straightfoward` should read `straightforward`
