+++ joss/paper.md
@@ -8,7 +8,7 @@ tags:
 authors:
   - name: Nicodemo Di Pasquale^[Corresponding author]
     orcid:  0000-0001-5676-8527
-    affiliation: 1
+    affiliation: "1,5"
   - name: Ruslan Davidchack
     orcid: 0000-0001-9418-5322
     affiliation: 4
@@ -24,6 +24,8 @@ affiliations:
    index: 3
  - name: Department of Mathematics, University of Leicester, United Kingdom
    index: 4
+ - name: *Current affiliation*: Department of Industrial Chemistry, University of Bologna, Italy
+   index: 5
 date: 06 Apr 2023
 bibliography: paper.bib
 ---
@@ -32,7 +34,7 @@ bibliography: paper.bib
 
 The numerical calculation of the properties of the interface between different materials or different phases of the same material has always been an important problem due to the insight it can bring into understanding the mechanisms underlying various interfacial phenomena, but also because of the challenges of measuring such properties experimentally.
 
-While the free energy of liquid-liquid and liquid-vapour interfaces is directly linked to the surface tension, which can be calculated from the anisotropy of the pressure tensor at the interface, the same approach cannot be used when a solid phase is present, since the interface introduces residual stress in the nearby solid which, in turn, changes the interfacial tension relative to the excess free energy [@DiPasquale2020shuttleworth]. As a result, a direct calculation of the free energy change during the formation of an interface from two separate bulk phases is required. While free energy calculation methods in general, and Thermodynamic Integration (TI) in particular, have been widely available in various molecular simulation packages, their use in conjunction with the complex geometry of an interfacial system is quite challenging and therefore much less common. Here we present a package for the calculation of the Interfacial Free Energy (IFE) of solid-vapour, solid-liquid and solid-solid interfaces for the widely used MD package LAMMPS (Large-Scale Atomic/Molecular Massively Parallel Simulator)[@Thompson2022].
+While the free energy of liquid-liquid and liquid-vapour interfaces is directly linked to the surface tension, which can be calculated from the anisotropy of the pressure tensor at the interface, the same approach cannot be used when a solid phase is present, since the interface introduces residual stresses in the nearby solid which, in turn, change the interfacial tension relative to the excess free energy [@DiPasquale2020shuttleworth]. As a result, a direct calculation of the free energy change during the formation of an interface from two separate bulk phases is required. While free energy calculation methods in general, and Thermodynamic Integration (TI) in particular, have been widely available in various molecular simulation packages, their use in conjunction with the complex geometry of an interfacial system is quite challenging and therefore much less common. Here we present a package for the calculation of the Interfacial Free Energy (IFE) of solid-vapour, solid-liquid and solid-solid interfaces for the widely used MD package LAMMPS (Large-Scale Atomic/Molecular Massively Parallel Simulator)[@Thompson2022].
 
 # Statement of need
 
@@ -49,13 +51,12 @@ Direct determination of the IFE from the reversible work of creating an interfac
 
 ![The thermodynamic path used to compute the interfacial free energy between phases $\alpha$ and $\beta$ with the cleaving method\label{fig:cleaving}](Fig/joss.png){ width=30% }
 
-A sketch of the cleaving path to obtain the IFE between two generic phases $\alpha$ and $\beta$ is shown in \autoref{fig:cleaving}. The reversible work in each step can be determined by different free energy computation methods, but those based on thermodynamic integration are the most straightfoward and provide accurate and direct results. With TI methods the free energy difference between two thermodynamic states connected by a transformation path is calculated by integrating the ensemble average of some configuration-dependent function (*e.g.*, the potential energy) with respect to a parameter defining the path from the initial to the final thermodynamic state.
[…]
