### 2024-01-29 #5886 @demonico85
> @demonico85 as AEiC I will now help to process this work for acceptance in JOSS. I have checked your repository, the archive link, this review, and the paper. Most seems in order, I only have the below minor points that require your attention.
>
>     * [ ]  `characterisation` is British English while `generalizable` is American English. Consider making the language choice consistent.
>
>     * [ ]  `straightfoward` should read `straightforward`

Thanks! I have corrected the mistakes. There is only one thing left, I can see that the paper is referenced as "Pasquale et al. (2024)." The correct spelling of my surname is "Di Pasquale" (i.e., "Di" is not a middle name), can you please help me to fix this?
