### 2024-02-06 a7f0f7c
New subs for type scaling
