+++ src/pair_lj_cut_tip4p_short.cpp
@@ -1,624 +1,0 @@
-// clang-format off
-/* ----------------------------------------------------------------------
-   LAMMPS - Large-scale Atomic/Molecular Massively Parallel Simulator
-   https://www.lammps.org/, Sandia National Laboratories
-   LAMMPS development team: developers@lammps.org
-
-   Copyright (2003) Sandia Corporation.  Under the terms of Contract
-   DE-AC04-94AL85000 with Sandia Corporation, the U.S. Government retains
-   certain rights in this software.  This software is distributed under
-   the GNU General Public License.
-
-   See the README file in the top-level LAMMPS directory.
-------------------------------------------------------------------------- */
-
-/* ----------------------------------------------------------------------
-   Contributing authors: Amalie Frischknecht and Ahmed Ismail (SNL)
-   simpler force assignment added by Rolf Isele-Holder (Aachen University)
-------------------------------------------------------------------------- */
-
-#include "pair_lj_cut_tip4p_short.h"
-
-#include "angle.h"
-#include "atom.h"
-#include "bond.h"
-#include "comm.h"
-#include "domain.h"
-#include "force.h"
-#include "neighbor.h"
-#include "neigh_list.h"
-#include "memory.h"
-#include "error.h"
-
-#include <cmath>
-#include <cstring>
-
-using namespace LAMMPS_NS;
-
-#define EWALD_F   1.12837917
-#define EWALD_P   0.3275911
-#define A1        0.254829592
-#define A2       -0.284496736
-#define A3        1.421413741
-#define A4       -1.453152027
-#define A5        1.061405429
-
-/* ---------------------------------------------------------------------- */
-
-PairLJCutTIP4PShort::PairLJCutTIP4PShort(LAMMPS *lmp) :
-  PairLJCutCoulLong(lmp)
-{
-  tip4pflag = 1;
-  ewaldflag = pppmflag = 1;  // for clarity, though inherited from parent class
-
-  single_enable = 0;
-  respa_enable = 0;
-  writedata = 1;
-
-  nmax = 0;
-  hneigh = nullptr;
-  newsite = nullptr;
-
-  // TIP4P cannot compute virial as F dot r
-  // due to finding bonded H atoms which are not near O atom
-
-  no_virial_fdotr_compute = 1;
-}
-
-/* ---------------------------------------------------------------------- */
-
-PairLJCutTIP4PShort::~PairLJCutTIP4PShort()
-{
-  memory->destroy(hneigh);
-  memory->destroy(newsite);
-}
-
-/* ---------------------------------------------------------------------- */
-
-void PairLJCutTIP4PShort::compute(int eflag, int vflag)
-{
-  int i,j,ii,jj,inum,jnum,itype,jtype,itable,key;
-  int n,vlist[6];
-  int iH1,iH2,jH1,jH2;
-  double qtmp,xtmp,ytmp,ztmp,delx,dely,delz,evdwl,ecoul;
-  double fraction,table;
-  double r,r2inv,r6inv,forcecoul,forcelj,cforce;
-  double factor_coul,factor_lj;
-  double grij,expm2,prefactor,t,erfc;
-  double fO[3],fH[3],fd[3],v[6];
-  double *x1,*x2,*xH1,*xH2;
-  int *ilist,*jlist,*numneigh,**firstneigh;
-  double rsq;
-
-  evdwl = ecoul = 0.0;
-  ev_init(eflag,vflag);
-
-  // reallocate hneigh & newsite if necessary
[…]
