### 2024-02-06 d6b77a1
New subs for type scaling
