+++ src/pair_lj_cut_tip4p_long_cleav_types.cpp
@@ -1,0 +1,882 @@
+// clang-format off
+/* ----------------------------------------------------------------------
+   LAMMPS - Large-scale Atomic/Molecular Massively Parallel Simulator
+   https://www.lammps.org/, Sandia National Laboratories
+   LAMMPS development team: developers@lammps.org
+
+   Copyright (2003) Sandia Corporation.  Under the terms of Contract
+   DE-AC04-94AL85000 with Sandia Corporation, the U.S. Government retains
+   certain rights in this software.  This software is distributed under
+   the GNU General Public License.
+
+   See the README file in the top-level LAMMPS directory.
+------------------------------------------------------------------------- */
+
+/* ----------------------------------------------------------------------
+   Contributing authors: Amalie Frischknecht and Ahmed Ismail (SNL)
+   simpler force assignment added by Rolf Isele-Holder (Aachen University)
+------------------------------------------------------------------------- */
+
+#include "pair_lj_cut_tip4p_long_cleav_types.h"
+
+#include "compute_chunk_atom.h"
+#include "compute_com_chunk.h"
+#include "angle.h"
+#include "atom.h"
+#include "bond.h"
+#include "comm.h"
+#include "domain.h"
+#include "force.h"
+#include "neighbor.h"
+#include "neigh_list.h"
+#include "memory.h"
+#include "error.h"
+#include "modify.h"
+
+#include <cmath>
+#include <cstring>
+
+using namespace LAMMPS_NS;
+
+#define EWALD_F   1.12837917
+#define EWALD_P   0.3275911
+#define A1        0.254829592
+#define A2       -0.284496736
+#define A3        1.421413741
+#define A4       -1.453152027
+#define A5        1.061405429
+
+/* ---------------------------------------------------------------------- */
+
+PairLJCutTIP4PLongCleavTypes::PairLJCutTIP4PLongCleavTypes(LAMMPS *lmp) :
+  PairLJCutCoulLong(lmp)
+{
+  tip4pflag = 1;
+  ewaldflag = pppmflag = 1;  // for clarity, though inherited from parent class
+
+  single_enable = 0;
+  respa_enable = 0;
+  writedata = 1;
+
+  nmax = 0;
+  hneigh = nullptr;
+  newsite = nullptr;
+
+  // TIP4P cannot compute virial as F dot r
+  // due to finding bonded H atoms which are not near O atom
+
+  no_virial_fdotr_compute = 1;
+  
+  ntypes=atom->ntypes;
+  natoms=atom->natoms; //array does start from zero
+
+
+
+  if(natoms == 0){
+      error->all(FLERR,"Pair style must be declared after box generation");
+  }
+  
+
+  pallocation = 0;
+  natoms=natoms+1;
+  dubtypes=ntypes*2;
+      
+  if(ntypes > 0){
+    nextra = dubtypes*dubtypes+1; // the zero-th position is not considered
+    pvector = new double[nextra];
+    pallocation = 1;
+    }
+
+   xprd = domain->xprd;
+   yprd = domain->yprd;
+   zprd = domain->zprd;
+   xy = domain->xy;
+   yz = domain->yz;
+   xz = domain->xz;
+  
+}
+
+/* ---------------------------------------------------------------------- */
+
+PairLJCutTIP4PLongCleavTypes::~PairLJCutTIP4PLongCleavTypes()
+{
[…]
