### 2024-02-06 b172ccb
Update fix_move_duplicate.cpp
