+++ src/pair_lj_cut_tip4p_long_cleav_types.cpp
@@ -87,13 +87,7 @@ PairLJCutTIP4PLongCleavTypes::PairLJCutTIP4PLongCleavTypes(LAMMPS *lmp) :
     pallocation = 1;
     }
 
-   xprd = domain->xprd;
-   yprd = domain->yprd;
-   zprd = domain->zprd;
-   xy = domain->xy;
-   yz = domain->yz;
-   xz = domain->xz;
-  
+
 }
 
 /* ---------------------------------------------------------------------- */
@@ -221,15 +215,12 @@ void PairLJCutTIP4PLongCleavTypes::compute(int eflag, int vflag)
 
   int t1, t2, scaling,m;  
   double flam,fDlam;
-
+    
 
   for(i=0; i<nextra ; i++){
     pvector[i] = 0.0;
     }
-    
-    
-
-    
+   
 
   // loop over neighbors of my atoms
 
@@ -304,7 +295,7 @@ void PairLJCutTIP4PLongCleavTypes::compute(int eflag, int vflag)
             }     
             
 
-        
+    
         
         forcelj = r6inv * (lj1[itype][jtype]*r6inv - lj2[itype][jtype]);
         forcelj *= flam * factor_lj * r2inv;
@@ -325,15 +316,15 @@ void PairLJCutTIP4PLongCleavTypes::compute(int eflag, int vflag)
           evdwl *= factor_lj;
           
 
-		  if(scaling){
+	   if(scaling){
                pvector[m] += fDlam * evdwl;
                evdwl *= flam;
                
-                       if (comm->me == 0)   printf("%d %d %d %d S %d \n LJ %f %f %f %f \n ",i,j,itype,jtype,scaling,evdwl,flam,fDlam,pvector[m] );  
+           //            if (comm->me == 0)   printf("%d %d %d %d S %d \n LJ %f %f %f %f \n ",i,j,itype,jtype,scaling,evdwl,flam,fDlam,pvector[m] );  
             }
             
             
-            if (comm->me == 0)   printf("%f\n",evdwl);
+ //           if (evdwl > 100)   printf("%d %d %d %d S %d \n LJ %f %f %f %f \n ",i,j,itype,jtype,scaling,evdwl,flam,fDlam,pvector[m] );  
         } else evdwl = 0.0;
 
         if (evflag) ev_tally(i,j,nlocal,newton_pair,
@@ -569,7 +560,7 @@ void PairLJCutTIP4PLongCleavTypes::compute(int eflag, int vflag)
 //fprintf(fp,"Scaled %d %d %f %f \n",tag[i],tag[j],ecoul,pvector[m]);
                pvector[m] += fDlam * ecoul;
                ecoul *= flam;
-                                      if (comm->me == 0)   printf("%d %d %d %d SC %d \n C %f %f %f %f \n ",i,j,itype,jtype,scaling,ecoul,flam,fDlam,pvector[m] );  
+//                                      if (comm->me == 0)   printf("%d %d %d %d SC %d \n C %f %f %f %f \n ",i,j,itype,jtype,scaling,ecoul,flam,fDlam,pvector[m] );  
 //fprintf(fp,"Scaled %f %f %f \n",fDlam,ecoul,pvector[m]);
             }
           } else ecoul = 0.0;
@@ -589,7 +580,7 @@ void PairLJCutTIP4PLongCleavTypes::compute(int eflag, int vflag)
 
 void PairLJCutTIP4PLongCleavTypes::settings(int narg, char **arg)
 {
[…]
