### 2024-02-07 #5886 @Kevin-Mattheus-Moerman
@openjournals/dev can you help with this author's issue:
> ... the paper is referenced as "Pasquale et al. (2024)." The correct spelling of my surname is "Di Pasquale" (i.e., "Di" is not a middle name), can you please help me to fix this?
