### 2024-02-07 1ca3979
Add a CITATION.cff following acceptance in JOSS
