### 2024-02-08 #5886 @xuanxu
> openjournals/dev can you help with this author's issue:
>
> > ... the paper is referenced as "Pasquale et al. (2024)." The correct spelling of my surname is "Di Pasquale" (i.e., "Di" is not a middle name), can you please help me to fix this?

We support several options to describe the parts of names in detail:
https://joss.readthedocs.io/en/latest/submitting.html#names
In this case probably setting `Di Pasquale` as the surname is enough, but there are other options like declaring a `non-dropping particle`.
