### 2024-02-08 #5886 @xuanxu
> is the accepted version going to be updated as well?

Sure!
@Kevin-Mattheus-Moerman: the `reaccept` command should do the trick
