### 2024-02-08 625932b
Update README.md with the JOSS reference
