### 2024-02-12 #5886 @Kevin-Mattheus-Moerman
@xuanxu I think the website still renders the paper with `Pasquale` rather than `Di Pasquale`:
![image](https://github.com/openjournals/joss-reviews/assets/8392709/74353c4b-d64c-41af-90d8-4edb9a85a0c5)

How can we fix that?
