+++ joss/paper.md
@@ -9,7 +9,7 @@ authors:
   - name:
        given-names: Nicodemo^[Corresponding author]
        non-dropping-particle: Di
-       surname: Pasquale
+       surname: Di Pasquale
     orcid:  0000-0001-5676-8527
     affiliation: "1,5"
   - name: Ruslan Davidchack
