+++ joss/paper.md
@@ -8,7 +8,6 @@ tags:
 authors:
   - name:
        given-names: Nicodemo^[Corresponding author]
-       non-dropping-particle: Di
        surname: Di Pasquale
     orcid:  0000-0001-5676-8527
     affiliation: "1,5"
