### 2024-02-14 #5886 @Kevin-Mattheus-Moerman
@demonico85 can your confirm you are now happy with the way the paper and citation are rendered?
