+++ src/pair_lj_cut_tip4p_long_cleav_opt.cpp
@@ -17,7 +17,7 @@
    simpler force assignment added by Rolf Isele-Holder (Aachen University)
 ------------------------------------------------------------------------- */
 
-#include "pair_lj_cut_tip4p_long_cleav.h"
+#include "pair_lj_cut_tip4p_long_cleav_opt.h"
 
 #include "compute_chunk_atom.h"
 #include "compute_com_chunk.h"
@@ -48,7 +48,7 @@ using namespace LAMMPS_NS;
 
 /* ---------------------------------------------------------------------- */
 
-PairLJCutTIP4PLongCleav::PairLJCutTIP4PLongCleav(LAMMPS *lmp) :
+PairLJCutTIP4PLongCleavOpt::PairLJCutTIP4PLongCleavOpt(LAMMPS *lmp) :
   PairLJCutCoulLong(lmp)
 {
   tip4pflag = 1;
@@ -71,6 +71,7 @@ PairLJCutTIP4PLongCleav::PairLJCutTIP4PLongCleav(LAMMPS *lmp) :
   natoms=atom->natoms; //array does start from zero
 
 
+  minlj_cut_off_sq = 1e10;
 
   if(natoms == 0){
       error->all(FLERR,"Pair style must be declared after box generation");
@@ -100,7 +101,7 @@ PairLJCutTIP4PLongCleav::PairLJCutTIP4PLongCleav(LAMMPS *lmp) :
 
 /* ---------------------------------------------------------------------- */
 
-PairLJCutTIP4PLongCleav::~PairLJCutTIP4PLongCleav()
+PairLJCutTIP4PLongCleavOpt::~PairLJCutTIP4PLongCleavOpt()
 {
   memory->destroy(hneigh);
   memory->destroy(newsite);
@@ -124,7 +125,7 @@ PairLJCutTIP4PLongCleav::~PairLJCutTIP4PLongCleav()
    allocate all arrays
 ------------------------------------------------------------------------- */
 
-void PairLJCutTIP4PLongCleav::allocate()
+void PairLJCutTIP4PLongCleavOpt::allocate()
 {
   allocated = 1;
   int n = atom->ntypes;
@@ -168,7 +169,21 @@ void PairLJCutTIP4PLongCleav::allocate()
 
 /* ---------------------------------------------------------------------- */
 
-void PairLJCutTIP4PLongCleav::compute(int eflag, int vflag)
+void PairLJCutTIP4PLongCleavOpt::compute(int eflag, int vflag)
+{
+
+  double cut_coulsqplus = (cut_coul+2.0*qdist) * (cut_coul+2.0*qdist);
+
+if(minlj_cut_off_sq < cut_coulsqplus) computeC(eflag,vflag);
+else computeLJ(eflag,vflag);
+
+
+
+}
+
+
+/* ---------------------------------------------------------------------- */
+void PairLJCutTIP4PLongCleavOpt::computeLJ(int eflag, int vflag)
 {
   int i,j,ii,jj,inum,jnum,itype,jtype,itable,key;
   int n,vlist[6];
@@ -416,6 +431,352 @@ void PairLJCutTIP4PLongCleav::compute(int eflag, int vflag)
             }
           }
 
+
+          cforce = flam*forcecoul * r2inv;
+          
+//printf("%d %d %d %d %f %f \n COUL %f %f %f\n ",i,j,itype,jtype,  powlambda[itype][jtype],powDlambda[itype][jtype],cut_coulsq,cut_coulsqplus,cforce); 
+
+          // if i,j are not O atoms, force is applied directly
+          // if i or j are O atoms, force is on fictitious atom & partitioned
+          // force partitioning due to Feenstra, J Comp Chem, 20, 786 (1999)
+          // f_f = fictitious force, fO = f_f (1 - 2 alpha), fH = alpha f_f
[…]
