### 2024-02-14 06f3285
Update pair_lj_cut_tip4p_long_cleav_opt.cpp
