### 2024-02-14 #5886 @demonico85
I can confirm that I am happy with the paper and the citation.

Thanks to all!
